add readme to p2p and move to pkg
