+++ node/setup.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/rollkit/rollkit/block"
 	"github.com/rollkit/rollkit/config"
-	"github.com/rollkit/rollkit/p2p"
+	"github.com/rollkit/rollkit/pkg/p2p"
 )
 
 const readHeaderTimeout = 10 * time.Second
